+++ scenedetect/detectors/content_detector.py
@@ -18,22 +18,17 @@
 # ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
+""":py:class:`ContentDetector` compares the difference in content between adjacent frames
+against a set threshold/score, which if exceeded, triggers a scene cut.
 
-""" ``scenedetect.detectors.content_detector`` Module
-
-This module implements the :py:class:`ContentDetector`, which compares the
-difference in content between adjacent frames against a set threshold/score,
-which if exceeded, triggers a scene cut.
-
-This detector is available from the command-line interface by using the
-`detect-content` command.
+This detector is available from the command-line as the `detect-content` command.
 """
 
-# Third-Party Library Imports
+from typing import List
+
 import numpy
 import cv2
 
-# PySceneDetect Library Imports
 from scenedetect.scene_detector import SceneDetector
 
 
@@ -49,10 +44,19 @@ class ContentDetector(SceneDetector):
     DELTA_H_KEY, DELTA_S_KEY, DELTA_V_KEY = ('delta_hue', 'delta_sat', 'delta_lum')
     METRIC_KEYS = [FRAME_SCORE_KEY, DELTA_H_KEY, DELTA_S_KEY, DELTA_V_KEY]
 
+    def __init__(self, threshold: float = 30.0, min_scene_len: int = 15, luma_only: bool = False):
+        """Construct a `ContentDetector`.
 
-    def __init__(self, threshold=30.0, min_scene_len=15, luma_only=False):
-        # type: (float, Union[int, FrameTimecode]) -> None
-        super(ContentDetector, self).__init__()
+        Arguments:
+            threshold: Threshold the average change in pixel intensity must exceed to trigger a cut.
+            min_scene_len: Once a cut is detected, this many frames must pass before a new one can
+                be added to the scene list.
+            luma_only: If True, only considers changes in the luminance channel of the video. The
+                default is False, which considers changes in hue, saturation, and luma.
+        """
+        #
+        #  type: (float, Union[int, FrameTimecode]) -> None
+        super().__init__()
         self.threshold = threshold
         # Minimum length of any given scene, in frames (int) or FrameTimecode
         self.min_scene_len = min_scene_len
@@ -61,52 +65,47 @@ class ContentDetector(SceneDetector):
         self.last_scene_cut = None
         self.last_hsv = None
 
-
     def get_metrics(self):
         return ContentDetector.METRIC_KEYS
 
-
     def is_processing_required(self, frame_num):
-        return self.stats_manager is None or (
-            not self.stats_manager.metrics_exist(frame_num, ContentDetector.METRIC_KEYS))
+        return self.stats_manager is None or (not self.stats_manager.metrics_exist(
+            frame_num, ContentDetector.METRIC_KEYS))
 
-
-    def calculate_frame_score(self, frame_num, curr_hsv, last_hsv):
-        # type: (int, List[numpy.ndarray], List[numpy.ndarray]) -> float
+    def _calculate_frame_score(self, frame_num: int, curr_hsv: List[numpy.ndarray],
+                               last_hsv: List[numpy.ndarray]) -> float:
         curr_hsv = [x.astype(numpy.int32) for x in curr_hsv]
         last_hsv = [x.astype(numpy.int32) for x in last_hsv]
         delta_hsv = [0, 0, 0, 0]
         for i in range(3):
             num_pixels = curr_hsv[i].shape[0] * curr_hsv[i].shape[1]
-            delta_hsv[i] = numpy.sum(
-                numpy.abs(curr_hsv[i] - last_hsv[i])) / float(num_pixels)
+            delta_hsv[i] = numpy.sum(numpy.abs(curr_hsv[i] - last_hsv[i])) / float(num_pixels)
 
         delta_hsv[3] = sum(delta_hsv[0:3]) / 3.0
         delta_h, delta_s, delta_v, delta_content = delta_hsv
 
         if self.stats_manager is not None:
-            self.stats_manager.set_metrics(frame_num, {
-                self.FRAME_SCORE_KEY: delta_content,
-                self.DELTA_H_KEY: delta_h,
-                self.DELTA_S_KEY: delta_s,
-                self.DELTA_V_KEY: delta_v})
+            self.stats_manager.set_metrics(
+                frame_num, {
+                    self.FRAME_SCORE_KEY: delta_content,
+                    self.DELTA_H_KEY: delta_h,
+                    self.DELTA_S_KEY: delta_s,
+                    self.DELTA_V_KEY: delta_v
+                })
         return delta_content if not self.luma_only else delta_v
 
-
-    def process_frame(self, frame_num, frame_img):
-        # type: (int, numpy.ndarray) -> List[int]
+    def process_frame(self, frame_num: int, frame_img: numpy.ndarray) -> List[int]:
         """ Similar to ThresholdDetector, but using the HSV colour space DIFFERENCE instead
         of single-frame RGB/grayscale intensity (thus cannot detect slow fades with this method).
 
         Arguments:
-            frame_num (int): Frame number of frame that is being passed.
-
-            frame_img (Optional[int]): Decoded frame image (numpy.ndarray) to perform scene
+            frame_num: Frame number of frame that is being passed.
+            frame_img: Decoded frame image (numpy.ndarray) to perform scene
                 detection on. Can be None *only* if the self.is_processing_required() method
                 (inhereted from the base SceneDetector class) returns True.
 
         Returns:
-            List[int]: List of frames where scene cuts have been detected. There may be 0
+            List of frames where scene cuts have been detected. There may be 0
             or more frames in the list, and not necessarily the same as frame_num.
         """
 
@@ -122,10 +121,10 @@ class ContentDetector(SceneDetector):
             # We obtain the change in average of HSV (frame_score), (h)ue only,
             # (s)aturation only, and (l)uminance only.  These are refered to in a statsfile
             # as their respective metric keys.
-            metric_key = (ContentDetector.DELTA_V_KEY if self.luma_only
-                          else ContentDetector.FRAME_SCORE_KEY)
-            if (self.stats_manager is not None and
-                    self.stats_manager.metrics_exist(frame_num, [metric_key])):
+            metric_key = (
+                ContentDetector.DELTA_V_KEY if self.luma_only else ContentDetector.FRAME_SCORE_KEY)
+            if (self.stats_manager is not None
+                    and self.stats_manager.metrics_exist(frame_num, [metric_key])):
                 frame_score = self.stats_manager.get_metrics(frame_num, [metric_key])[0]
             else:
                 curr_hsv = cv2.split(cv2.cvtColor(frame_img, cv2.COLOR_BGR2HSV))
@@ -133,14 +132,14 @@ class ContentDetector(SceneDetector):
                 if not last_hsv:
                     last_hsv = cv2.split(cv2.cvtColor(self.last_frame, cv2.COLOR_BGR2HSV))
 
-                frame_score = self.calculate_frame_score(frame_num, curr_hsv, last_hsv)
+                frame_score = self._calculate_frame_score(frame_num, curr_hsv, last_hsv)
 
                 self.last_hsv = curr_hsv
 
             # We consider any frame over the threshold a new scene, but only if
             # the minimum scene length has been reached (otherwise it is ignored).
             if frame_score >= self.threshold and (
-                    (frame_num - self.last_scene_cut) >= self.min_scene_len):
+                (frame_num - self.last_scene_cut) >= self.min_scene_len):
                 cut_list.append(frame_num)
                 self.last_scene_cut = frame_num
 
@@ -149,14 +148,13 @@ class ContentDetector(SceneDetector):
 
         # If we have the next frame computed, don't copy the current frame
         # into last_frame since we won't use it on the next call anyways.
-        if (self.stats_manager is not None and
-                self.stats_manager.metrics_exist(frame_num+1, self.get_metrics())):
+        if (self.stats_manager is not None
+                and self.stats_manager.metrics_exist(frame_num + 1, self.get_metrics())):
             self.last_frame = _unused
         else:
             self.last_frame = frame_img.copy()
 
         return cut_list
-
 
     #def post_process(self, frame_num):
     #    """ TODO: Based on the parameters passed to the ContentDetector constructor,