+++ scenedetect/detectors/threshold_detector.py
@@ -18,39 +18,36 @@
 # ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
+""":py:class:`ThresholdDetector` uses a set intensity as a threshold to detect cuts, which
+are triggered when the average pixel intensity exceeds or falls below this threshold.
 
-""" ``scenedetect.detectors.threshold_detector`` Module
-
-This module implements the :py:class:`ThresholdDetector`, which uses a set intensity
-level as a threshold, to detect cuts when the average frame intensity exceeds or falls
-below this threshold.
-
-This detector is available from the command-line interface by using the
-`detect-threshold` command.
+This detector is available from the command-line as the `detect-threshold` command.
 """
 
-# Third-Party Library Imports
+from typing import List
+
 import numpy
 
-# PySceneDetect Library Imports
 from scenedetect.scene_detector import SceneDetector
-
 
 ##
 ## ThresholdDetector Helper Functions
 ##
 
-def compute_frame_average(frame):
+
+def compute_frame_average(frame: numpy.ndarray) -> float:
     """Computes the average pixel value/intensity for all pixels in a frame.
 
     The value is computed by adding up the 8-bit R, G, and B values for
     each pixel, and dividing by the number of pixels multiplied by 3.
 
+    Arguments:
+        frame: Frame representing the RGB pixels to average.
+
     Returns:
-        Floating point value representing average pixel intensity.
+        Average pixel intensity across all 3 channels of `frame`
     """
-    num_pixel_values = float(
-        frame.shape[0] * frame.shape[1] * frame.shape[2])
+    num_pixel_values = float(frame.shape[0] * frame.shape[1] * frame.shape[2])
     avg_pixel_value = numpy.sum(frame[:, :, :]) / num_pixel_values
     return avg_pixel_value
 
@@ -59,35 +56,36 @@ def compute_frame_average(frame):
 ## ThresholdDetector Class Implementation
 ##
 
+
 class ThresholdDetector(SceneDetector):
     """Detects fast cuts/slow fades in from and out to a given threshold level.
 
     Detects both fast cuts and slow fades so long as an appropriate threshold
     is chosen (especially taking into account the minimum grey/black level).
-
-    Attributes:
-        threshold:  8-bit intensity value that each pixel value (R, G, and B)
-            must be <= to in order to trigger a fade in/out.
-        min_scene_len:  FrameTimecode object or integer greater than 0 of the
-            minimum length, in frames, of a scene (or subsequent scene cut).
-        fade_bias:  Float between -1.0 and +1.0 representing the percentage of
-            timecode skew for the start of a scene (-1.0 causing a cut at the
-            fade-to-black, 0.0 in the middle, and +1.0 causing the cut to be
-            right at the position where the threshold is passed).
-        add_final_scene:  Boolean indicating if the video ends on a fade-out to
-            generate an additional scene at this timecode.
-        block_size:  Number of rows in the image to sum per iteration (can be
-            tuned to increase performance in some cases; should be computed
-            programmatically in the future).
     """
 
     THRESHOLD_VALUE_KEY = 'delta_rgb'
 
-    def __init__(self, threshold=12, min_scene_len=15, fade_bias=0.0,
-                 add_final_scene=False, block_size=8):
-        """Initializes threshold-based scene detector object."""
+    def __init__(self,
+                 threshold: float = 12,
+                 min_scene_len: int = 15,
+                 fade_bias: float = 0.0,
+                 add_final_scene: bool = False):
+        """Construct a `ThresholdDetector`.
 
-        super(ThresholdDetector, self).__init__()
+        Arguments:
+            threshold:  8-bit intensity value that each pixel value (R, G, and B)
+                must be <= to in order to trigger a fade in/out.
+            min_scene_len:  FrameTimecode object or integer greater than 0 of the
+                minimum length, in frames, of a scene (or subsequent scene cut).
+            fade_bias:  Float between -1.0 and +1.0 representing the percentage of
+                timecode skew for the start of a scene (-1.0 causing a cut at the
+                fade-to-black, 0.0 in the middle, and +1.0 causing the cut to be
+                right at the position where the threshold is passed).
+            add_final_scene:  Boolean indicating if the video ends on a fade-out to
+                generate an additional scene at this timecode.
+        """
+        super().__init__()
         self.threshold = int(threshold)
         self.fade_bias = fade_bias
         self.min_scene_len = min_scene_len
@@ -98,19 +96,15 @@ class ThresholdDetector(SceneDetector):
         self.add_final_scene = add_final_scene
         # Where the last fade (threshold crossing) was detected.
         self.last_fade = {
-            'frame': 0,         # frame number where the last detected fade is
-            'type': None        # type of fade, can be either 'in' or 'out'
+            'frame': 0,  # frame number where the last detected fade is
+            'type': None  # type of fade, can be either 'in' or 'out'
         }
-        self.block_size = block_size
         self._metric_keys = [ThresholdDetector.THRESHOLD_VALUE_KEY]
 
-
-    def get_metrics(self):
+    def get_metrics(self) -> List[str]:
         return self._metric_keys
 
-
-    def process_frame(self, frame_num, frame_img):
-        # type: (int, Optional[numpy.ndarray]) -> List[int]
+    def process_frame(self, frame_num: int, frame_img: numpy.ndarray) -> List[int]:
         """
         Args:
             frame_num (int): Frame number of frame that is being passed.
@@ -137,15 +131,13 @@ class ThresholdDetector(SceneDetector):
         # less than or equal to the threshold; however, since this differs on
         # user-supplied values, we supply the average pixel intensity as this
         # frame metric instead (to assist with manually selecting a threshold)
-        if (self.stats_manager is not None) and (
-                self.stats_manager.metrics_exist(frame_num, self._metric_keys)):
-            frame_avg = self.stats_manager.get_metrics(
-                frame_num, self._metric_keys)[0]
+        if (self.stats_manager is not None) and (self.stats_manager.metrics_exist(
+                frame_num, self._metric_keys)):
+            frame_avg = self.stats_manager.get_metrics(frame_num, self._metric_keys)[0]
         else:
             frame_avg = compute_frame_average(frame_img)
             if self.stats_manager is not None:
-                self.stats_manager.set_metrics(
-                    frame_num, {self._metric_keys[0]: frame_avg})
+                self.stats_manager.set_metrics(frame_num, {self._metric_keys[0]: frame_avg})
 
         if self.processed_frame:
             if self.last_fade['type'] == 'in' and frame_avg < self.threshold:
@@ -158,8 +150,8 @@ class ThresholdDetector(SceneDetector):
                     # Just faded into a new scene, compute timecode for the scene
                     # split based on the fade bias.
                     f_out = self.last_fade['frame']
-                    f_split = int((frame_num + f_out +
-                                   int(self.fade_bias * (frame_num - f_out))) / 2)
+                    f_split = int(
+                        (frame_num + f_out + int(self.fade_bias * (frame_num - f_out))) / 2)
                     cut_list.append(f_split)
                     self.last_scene_cut = frame_num
                 self.last_fade['type'] = 'in'
@@ -173,8 +165,7 @@ class ThresholdDetector(SceneDetector):
         self.processed_frame = True
         return cut_list
 
-
-    def post_process(self, frame_num):
+    def post_process(self, frame_num: int):
         """Writes a final scene cut if the last detected fade was a fade-out.
 
         Only writes the scene cut if add_final_scene is true, and the last fade
@@ -188,7 +179,7 @@ class ThresholdDetector(SceneDetector):
         # fade-outs, as a scene cut is already added when a fade-in is found.
         cut_times = []
         if self.last_fade['type'] == 'out' and self.add_final_scene and (
-                (self.last_scene_cut is None and frame_num >= self.min_scene_len) or
-                (frame_num - self.last_scene_cut) >= self.min_scene_len):
+            (self.last_scene_cut is None and frame_num >= self.min_scene_len) or
+            (frame_num - self.last_scene_cut) >= self.min_scene_len):
             cut_times.append(self.last_fade['frame'])
         return cut_times