+++ tests/test_stats_manager.py
@@ -234,7 +234,7 @@ def test_save_load_from_video(test_video_file):
     scene_manager.auto_downscale = True
     scene_manager.detect_scenes(video, duration=duration)
 
-    stats_manager.save_to_csv(TEST_STATS_FILES[0], base_timecode=base_timecode)
+    stats_manager.save_to_csv(path=TEST_STATS_FILES[0], base_timecode=base_timecode)
 
     stats_manager_new = StatsManager()